Extracted types into seperate file and removed unwanted error handling
